requirements.txt
flask
flask-cors
textblob
textstat
keybert
scikit-learn
sentence-transformers
openai
python-dotenv
